backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
beanie==1.26.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
email-validator==2.2.0
requests==2.32.3
motor==3.6.0
pymongo>=4.9.0
pandas==2.2.2

celery==5.4.0
redis==5.0.7

pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
